add support for collection verify
This commit adds the minimal required logic to process
instructions related to collection verification. Right now,
the only setting where that's needed is `SetAndVerifyCollection`
because collection information is already available for the
individual (un)verify operations.

This commit also updated the key index when parsing the creator
account pubkey to 5 (which is the correct 0-based index when
analyzing the accounts struct for those particular operations).
The previous strange behaviour is no longer present since the
validator container has been updated to a more recent version of
the plugin repo, and things are working as expected right now.

Signed-off-by: Alexandru Agache <alexandruag@gmail.com>